Use size_t for variables holding strlen() return values
CVS-ID: concat.c 1.2
CVS-ID: pam_passwdqc.c 1.24
CVS-ID: passwdqc_check.c 1.31
CVS-ID: pwqcheck.c 1.15

+++ concat.c
@@ -27,7 +27,7 @@ char *concat(const char *s1, ...)
 	va_list args;
 	const char *s;
 	char *p, *result;
-	unsigned long l, m, n;
+	size_t l, m, n;
 
 	m = n = strlen(s1);
 	va_start(args, s1);
